chore(code-refactor) Refactor routes and controllers This chore refactors the routes and controllers codes, also implemented good use of middlewares
