+++ backend-model/DiaryEntryModel.js
