--- backend-validation/validation.js
@@ -1,11 +1,0 @@
-import Joi from 'joi';
-
-export default (entry) => {
-  const schema = Joi.object().keys({
-    title: Joi.string().required(),
-    category: Joi.string().required(),
-    sub_category: Joi.string().required(),
-    content: Joi.string().required(),
-  });
-  return Joi.validate(entry, schema);
-};
